Basic Enemy Generator and Enemy Logic
Added very basic enemy generator that will spawn enemies along a preset path.  The enemy will follow the path at a set speed until colliding with the base at which point the enemy object will be destroyed.  This is just a baseline and I will be adding more complexity but we should be able to use this to proceed with other logic.